requirements.txt
# CareerBot — Dependencies (V2 - Gemini Embedding RAG)
# Railway Docker image target: <200MB (was ~600MB)

# Telegram bot (async, with job queue for triggers)
python-telegram-bot[job-queue]>=21.0

# Gemini AI (new SDK) - used for chat, classification, AND embeddings
google-genai>=1.0.0

# PDF generation
reportlab>=4.0.0

# PDF/DOCX reading (for career document ingestion)
pypdf>=4.0.0
python-docx>=1.0.0

# HTTP + env + Scraping
httpx>=0.27.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0

# Pipeline Dependencies (Job Search)
python-jobspy>=1.1.0
pandas>=2.0.0

# NOTE: chromadb and sentence-transformers REMOVED (were ~520MB)
# RAG now uses Gemini text-embedding-004 API + JSON vector store (~50KB)

# PostgreSQL (replaces SQLite for multi-tenant cloud DB)
psycopg2-binary>=2.9.9

# BYOK key encryption (AES-256 Fernet)
cryptography>=42.0.0

# Vercel Webhook / Serverless API
fastapi>=0.110.0
uvicorn>=0.29.0
